feat(typescript-react-apollo): support useBackgroundQuery and useLoadableQuery
Implement `useBackgroundQuery` and `useLoadableQuery` hooks for React
Apollo. These APIs are added in 3.8.0 and 3.9.0 respectively.

Suspense support in Apollo Client provides 5 new hooks as described in
https://www.apollographql.com/docs/react/data/suspense/.
`useSuspenseQuery` support was added in
https://github.com/dotansimha/graphql-code-generator-community/pull/434.
Among other 4 hooks, `useBackgroundQuery` and `useLoadableQuery` are
the ones that take a GraphQL document as an argument like other existing
hooks. The support for these hooks are added in this change.

Other hooks, `useQueryRefHanders` and `useReadQuery` do not take a
GraphQL document as an argument, so they do not need to be added as the
generated code. Both take the return values of other hooks as an
argument, and the types are inferred from it.

Follow up to
https://github.com/dotansimha/graphql-code-generator-community/issues/388.
Fixes https://github.com/dotansimha/graphql-code-generator-community/issues/477.